hackthon_bvp_3.0_django/requirements.txt
Django==2.2.6
gunicorn==19.9.0
pytz==2019.3
sqlparse==0.3.0
firebase_admin
google.cloud
numpy